requirements.txt
requests>=2.1.0
pycrypto>=2.6.1
